+++ tfjs-core/src/backends/kernel_impls.ts
@@ -16,5 +16,4 @@
  */
 
 export {nonMaxSuppressionV3Impl, nonMaxSuppressionV4Impl, nonMaxSuppressionV5Impl} from './non_max_suppression_impl';
-export {split} from './split_shared';
 export {whereImpl} from './where_impl';
